requirements.txt
#ala
Flask==3.0.0 
#doua
opencv-python-headless==4.8.1.78
numpy>=1.26.0
tensorflow>=2.14.0
torch>=2.1.0
torchvision>=0.16.0
pillow>=10.0.0
matplotlib>=3.7.2
scikit-learn>=1.3.0
tqdm>=4.65.0 
#ranim
ultralytics>=8.3.0
easyocr>=1.7.2

#hazem
face_recognition==1.3.0
#aziz
fastai==2.7.19 
librosa
